backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# RAG and Vector Database
chromadb==1.2.2
openai==2.6.1

# Text Processing
nltk==3.9.2

# Utilities
httpx==0.27.0
python-multipart==0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# CORS (built into FastAPI)

